Completed three (nose) tests in tests/: standard publish demo, user config with import of publish defaults, user config with no import of publish defaults.

+++ tests/copy_new_reference_data.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 cd reference_data
 for filename in `/bin/ls`; do
-    cp ../$filename $filename
+    if [ -f ../$filename ]; then
+        cp ../$filename $filename
+    fi
 done
